+++ OpenOrchestrator/database/triggers.py
@@ -24,6 +24,8 @@ class TriggerStatus(enum.Enum):
     DONE = "Done"
     PAUSED = "Paused"
     PAUSING = "Pausing"
+    KILLING = "Killing"
+    KILLED = "Killed"
 
 
 class TriggerType(enum.Enum):